Enhancement: Extract DispatchWebhooksCommand
